backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
djoser==2.2.0
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
requests==2.31.0
django-import-export==3.2.0
drf-extra-fields==3.5.0
django-cors-headers==4.2.0
django-colorfield==0.9.0